+++ backend/Resume-And-Portfolio/src/main/java/com/example/resumeandportfolio/filter/LoginFilter.java
@@ -1,7 +1,7 @@
 package com.example.resumeandportfolio.filter;
 
 import com.example.resumeandportfolio.model.entity.user.CustomUserDetails;
-import com.example.resumeandportfolio.service.global.RefreshTokenService;
+import com.example.resumeandportfolio.service.user.RefreshTokenService;
 import com.example.resumeandportfolio.util.jwt.JwtUtil;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.http.Cookie;
