+++ backend/Resume-And-Portfolio/src/test/java/com/example/resumeandportfolio/filter/CustomLogoutFilterTest.java
@@ -1,6 +1,6 @@
 package com.example.resumeandportfolio.filter;
 
-import com.example.resumeandportfolio.service.global.RefreshTokenService;
+import com.example.resumeandportfolio.service.user.RefreshTokenService;
 import com.example.resumeandportfolio.util.jwt.JwtUtil;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.http.Cookie;
